Add webpack-dev-server and start script
This commit adds webpack-dev-server as a development dependency and
adds a new "start" script to the package.json file. The "start" script
uses webpack-dev-server to serve the application, opening a new browser
window when the server starts.

Additionally, the webpack configuration file is updated to specify the
"./dist" directory as the static directory for the dev server. This
allows webpack-dev-server to serve the bundled application files for
development testing.

These changes improve the development experience by enabling hot
reloading and eliminating the need to manually rebuild and refresh the
application during development.